move PathAssignError to core to remove circularity around A, raise PathAccessError on s magic failure, close up a risky exception message hole in pathaccesserror.
